% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{list_routes}
\alias{list_routes}
\title{Route list}
\usage{
list_routes(agency)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}
}
\value{
route_list
}
\description{
Route list
}

