% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{nb_list_routes}
\alias{nb_list_routes}
\title{Route list}
\usage{
nb_list_routes(agency)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}
}
\value{
route_list
}
\description{
List of routes for a  given bus agency
}

