% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{nb_messages}
\alias{nb_messages}
\title{Agency Messages}
\usage{
nb_messages(agency, route)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}

\item{route}{The name of the route. Can get from \code{route_list()} function}
}
\value{
messages
}
\description{
Get the messages for selected routes
}

