% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{nb_vehicle_locations}
\alias{nb_vehicle_locations}
\title{Vehicle Location}
\usage{
nb_vehicle_locations(agency, route, time = 0)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}

\item{route}{The name of the route. Can get from \code{route_list()} function}

\item{time}{The unix time in msec. Default of t = 0 returns the most recent 15 minutes of data.}
}
\value{
vehicle_location
}
\description{
Get the location of a vehicle
}

