% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBus.R
\docType{package}
\name{NextBus}
\alias{NextBus}
\alias{NextBus-package}
\title{NextBus: Wrapper for the NextBus API}
\description{
This package is a connection to the NextBus API.
}
\section{NextBus functions}{

The NextBus functions
}

