% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{nextbus_api}
\alias{nextbus_api}
\title{NextBus API}
\usage{
nextbus_api(param)
}
\arguments{
\item{param}{a list of parameters for the GET query. Built as part of each of the functions.}
}
\value{
xml_parse
}
\description{
NextBus API
}

