% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{nextbus_url}
\alias{nextbus_url}
\title{NextBus API}
\usage{
nextbus_url(base_url = NULL)
}
\arguments{
\item{base_url}{The base url for the NextBus webservice}
}
\value{
nextbus_url
}
\description{
Base URL function
}

