% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{predictions}
\alias{predictions}
\title{Predictions}
\usage{
predictions(agency, route, stop)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}

\item{route}{The name of the route. Can get from \code{route_list()} function}

\item{stop}{The name of the stop. Can get from \code{route_config()} function.}
}
\value{
predictions
}
\description{
Combine the two predictions for stops
}

