% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBus.R
\docType{package}
\name{rnextbus}
\alias{rnextbus}
\alias{rnextbus-package}
\title{rnextbus: Wrapper for the NextBus API}
\description{
This package is a connection to the NextBus API.
}
\section{NextBus functions}{

The NextBus functions
nb_list_agencies
nb_list_routes
nb_route_config
nb_schedule
nb_messages
nb_vehicle_locations
}

