% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{route_config}
\alias{route_config}
\title{Route Configuration}
\usage{
route_config(agency, route)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}

\item{route}{The name of the route. Can get from \code{route_list()} function}
}
\value{
route_config
}
\description{
Route Configuration
}

