% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextBusAPI.R
\name{schedule}
\alias{schedule}
\title{Schedule}
\usage{
schedule(agency, route)
}
\arguments{
\item{agency}{The tag name of the bus agency. Can get from \code{list_agencies()} function}

\item{route}{The name of the route. Can get from \code{route_list()} function}
}
\value{
schedule_df
}
\description{
Get the schedule for a particular route
}

